2024/09/23 PriceService unit test

+++ common-library/src/main/java/com/jack/common/constants/ApplicationConstants.java
@@ -12,6 +12,6 @@ public class ApplicationConstants {
     public static final int DEFAULT_PAGE_NUMBER = 0;
     public static final int DEFAULT_PAGE_SIZE = 20;
 
-    public static final BigDecimal INITIAL_PRICE = BigDecimal.valueOf(100);
+    public static final double INITIAL_PRICE = 100.00;
     public static final String BTC_PRICE_KEY = "BTC_CURRENT_PRICE";
 }
